on-device/requirements.txt
# PyTorch
wget https://developer.download.nvidia.cn/compute/redist/jp/v512/pytorch/torch-2.1.0a0+41361538.nv23.06-cp38-cp38-linux_aarch64.whl
pip install torch-2.1.0a0+41361538.nv23.06-cp38-cp38-linux_aarch64.whl

# TorchVision 
sudo apt-get update
sudo apt-get install -y libjpeg-dev zlib1g-dev libavcodec-dev libavformat-dev libswscale-dev libpng-dev
sudo apt-get install libgtk2.0-dev pkg-config
sudo apt-get install libcanberra-gtk-module libcanberra-gtk3-module
git clone --branch v0.16.1 https://github.com/pytorch/vision torchvision
cd torchvision
python setup.py install

# List of Required Libraries
annotated_types==0.7.0
cffi==1.15.1
datasets==2.14.6
docstring_parser==0.15
edge_tts
face_recognition==1.3.0
speechrecognition
fastapi==0.103.0
GitPython==3.1.31
googletrans==4.0.0rc1
huggingface_hub==0.20.3
matplotlib==3.7.2
mistral_common==1.2.0
numpy==1.24.4
opencv-python==4.8.0.76
pandas==2.0.3
Pillow==9.5.0
pydantic==1.10.14
pytest==7.4.2
PyYAML==6.0
Requests==2.31.0
safetensors==0.4.2
scikit_learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
tabulate==0.9.0
tqdm==4.65.0
transformers==4.35.2
typer==0.9.0
typing_extensions==4.7.1
uvicorn==0.23.2
wget==3.2
python-multipart